Update product_template_attribute_line.py
When filtering as the following:
            archived_ptal = self.search([
                ('active', '=', False),
                ('product_tmpl_id', '=', vals.pop('product_tmpl_id', 0)),
                ('attribute_id', '=', vals.pop('attribute_id', 0)),
            ], limit=1)
We have the following behaviour: when the attribute line is archived, then it becomes active but it was active, a new product_template_attribute_line will be created. By removing active = false in the filter, we will have a better behaviour: when the product_template_attribute_line is archived (active = false), then it becomes active but when the product_template_attribute_line still exists, it will be completed with the product_template_attribute_value instead of creating a new product_template_attribute_line record.